requirements.txt
# Core dependencies for AI Video Analysis Platform
yt-dlp>=2023.12.30
openai-whisper>=20231117
requests>=2.31.0
python-dotenv>=1.0.0

# AI and ML dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0

# CLI and interface
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# Data processing
json5>=0.9.0
jsonschema>=4.19.0

# Audio/Video processing
ffmpeg-python>=0.2.0

# Utilities
pathlib2>=2.3.0
hashlib
datetime
typing-extensions>=4.8.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 